+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix table component alignment and styling
 - Fix toggle button styling
 - Miscellanous code improvements
+- Improve container index table
 
 # v0.7.1
 - Add shading to table component